backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi>=0.135.1",
    "motor>=3.7.1",
    "pydantic[email]>=2.12.5",
    "pyjwt>=2.11.0",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "uvicorn[standard]>=0.41.0",
]
